rm old path
